Use minimal Rust interpreter by default
old_interpreter feature can be used to switch back to the previous Rust
interpreter. Add (interpret atom type space) to the old interpreter for
compatibility. Remove previous minimal MeTTa interpreter. Fix unit tests
which depends on a bare interpreter.
