+++ lib/Cargo.toml
@@ -27,12 +27,10 @@ path = "src/lib.rs"
 crate-type = ["lib"]
 
 [features]
-default = ["minimal", "minimal_rust", "pkg_mgmt"]
+default = ["pkg_mgmt"]
 # Add one of the features below into default list to enable.
 # See https://doc.rust-lang.org/cargo/reference/features.html#the-features-section
-old_interpreter = [] # enables minimal MeTTa interpreter
-minimal = [] # enables minimal MeTTa interpreter
-minimal_rust = [] # enables minimal MeTTa interpreter with Rust interpreter
+old_interpreter = [] # enables old Rust interpreter
 variable_operation = [] # enables evaluation of the expressions which have
                         # a variable on the first position
 git = ["git2", "pkg_mgmt"]
