+++ lib/src/metta/runner/mod.rs
@@ -87,16 +87,11 @@ pub use environment::{Environment, EnvBuilder};
 
 #[macro_use]
 pub mod stdlib;
-#[cfg(not(feature = "minimal"))]
 use super::interpreter::{interpret, interpret_init, interpret_step, InterpreterState};
 
-#[cfg(feature = "minimal")]
+#[cfg(not(feature = "old_interpreter"))]
 pub mod stdlib_minimal;
-#[cfg(all(feature = "minimal", not(feature = "minimal_rust")))]
-use super::interpreter_minimal::{interpret, interpret_init, interpret_step, InterpreterState};
-#[cfg(all(feature = "minimal", feature = "minimal_rust"))]
-use super::interpreter_minimal_rust::{interpret, interpret_init, interpret_step, InterpreterState};
-#[cfg(feature = "minimal")]
+#[cfg(not(feature = "old_interpreter"))]
 use stdlib_minimal::*;
 
 use stdlib::CoreLibLoader;
@@ -444,10 +439,10 @@ impl Metta {
         state.run_to_completion()
     }
 
-    // TODO: this method is deprecated and should be removed after switching
+    // TODO: MINIMAL this method is deprecated and should be removed after switching
     // to the minimal MeTTa
     pub fn evaluate_atom(&self, atom: Atom) -> Result<Vec<Atom>, String> {
-        #[cfg(feature = "minimal")]
+        #[cfg(not(feature = "old_interpreter"))]
         let atom = if is_bare_minimal_interpreter(self) {
             atom
         } else {
@@ -1068,7 +1063,7 @@ impl<'input> RunContext<'_, '_, 'input> {
                                 let type_err_exp = Atom::expr([ERROR_SYMBOL, atom, BAD_TYPE_SYMBOL]);
                                 self.i_wrapper.interpreter_state = Some(InterpreterState::new_finished(self.module().space().clone(), vec![type_err_exp]));
                             } else {
-                                #[cfg(feature = "minimal")]
+                                #[cfg(not(feature = "old_interpreter"))]
                                 let atom = if is_bare_minimal_interpreter(self.metta) {
                                     atom
                                 } else {
@@ -1094,7 +1089,7 @@ impl<'input> RunContext<'_, '_, 'input> {
 
 }
 
-#[cfg(feature = "minimal")]
+#[cfg(not(feature = "old_interpreter"))]
 fn is_bare_minimal_interpreter(metta: &Metta) -> bool {
     metta.get_setting_string("interpreter") == Some("bare-minimal".into())
 }
@@ -1187,7 +1182,7 @@ impl<'i> InputStream<'i> {
     }
 }
 
-#[cfg(feature = "minimal")]
+#[cfg(not(feature = "old_interpreter"))]
 fn wrap_atom_by_metta_interpreter(space: DynSpace, atom: Atom) -> Atom {
     let space = Atom::gnd(space);
     let interpret = Atom::expr([Atom::sym("interpret"), atom, ATOM_TYPE_UNDEFINED, space]);
