+++ lib/src/metta/runner/modules/mod.rs
@@ -7,14 +7,12 @@ use crate::metta::runner::*;
 
 use regex::Regex;
 
-#[cfg(not(feature = "minimal"))]
+#[cfg(feature = "old_interpreter")]
 use super::stdlib::*;
 
-#[cfg(all(feature = "minimal", not(feature = "minimal_rust")))]
+#[cfg(not(feature = "old_interpreter"))]
 use super::interpreter_minimal::interpret;
-#[cfg(all(feature = "minimal", feature = "minimal_rust"))]
-use super::interpreter_minimal_rust::interpret;
-#[cfg(feature = "minimal")]
+#[cfg(not(feature = "old_interpreter"))]
 use super::stdlib_minimal::*;
 
 mod mod_names;
