+++ repl/Cargo.toml
@@ -23,5 +23,5 @@ path = "src/main.rs"
 default = ["python"]
 no_python = ["hyperon"]
 python = ["pyo3", "pep440_rs"]
-minimal = ["hyperon/minimal", "no_python"] #WARNING: The interpreter belonging to the hyperon python module will be used if hyperon is run through python
+old_interpreter = ["hyperon/old_interpreter"]
 git = ["hyperon/git"]
